+++ src/main/groovy/org/codenarc/util/ImportUtil.groovy
@@ -113,18 +113,27 @@ class ImportUtil {
         }
     }
 
-    static List getImportsSortedByLineNumber(SourceCode sourceCode) {
+    static List<ImportNode> getAllImports(SourceCode sourceCode) {
+        def ast = sourceCode.ast
+        if (!ast) {
+            return []
+        }
+        def staticImports = ast.staticImports.values() + ast.staticStarImports.values()
+        return ast.imports + ast.starImports + staticImports
+    }
+
+    static List<ImportNode> getImportsSortedByLineNumber(SourceCode sourceCode) {
         def staticImports = sourceCode.ast.staticImports.values() + sourceCode.ast.staticStarImports.values()
         def allImports = sourceCode.ast.imports + sourceCode.ast.starImports + staticImports
         return sortImportsByLineNumber(allImports, sourceCode)
     }
 
-    static List getNonStaticImportsSortedByLineNumber(SourceCode sourceCode) {
+    static List<ImportNode> getNonStaticImportsSortedByLineNumber(SourceCode sourceCode) {
         def allImports = sourceCode.ast.imports + sourceCode.ast.starImports
         return sortImportsByLineNumber(allImports, sourceCode)
     }
 
-    private static List sortImportsByLineNumber(List imports, SourceCode sourceCode) {
+    private static List<ImportNode> sortImportsByLineNumber(List<ImportNode> imports, SourceCode sourceCode) {
         imports.sort { importNode ->
             def importInfo = ImportUtil.sourceLineAndNumberForImport(sourceCode, importNode)
             importInfo.lineNumber